requirements.txt
torch>=2.0.0
transformers>=4.30.0
datasets>=2.12.0
soundfile>=0.12.1
librosa>=0.10.0
sounddevice>=0.4.6
requests>=2.31.0
tqdm>=4.65.0
nemo_toolkit[all]>=1.20.0 